+++ tests/nn/data_parallel/test_fsdp.py
@@ -33,8 +33,6 @@ from fairscale.utils.testing import (
 if torch_version() >= (1, 8, 0):
     from fairscale.optim.grad_scaler import ShardedGradScaler
 
-import tempfile
-
 # How to use remote-pdb: https://gist.github.com/sshleifer/9d43351957179c13606e015b072927d4
 # All helper functions called by spawn must be either @classmethod, @staticmethod
 
@@ -84,6 +82,9 @@ class DistributedTest(unittest.TestCase):
 
     @staticmethod
     def get_wrapped_model(group, cuda_first=False, config={}, **model_kwargs) -> FullyShardedDataParallel:
+        if config["offload_config"]:
+            return FullyShardedDataParallel(TransformerWithSharedParams(group, **model_kwargs), group, **config)
+
         if cuda_first:
             model = FullyShardedDataParallel(TransformerWithSharedParams(group, **model_kwargs).cuda(), group, **config)
         else:
@@ -151,6 +152,13 @@ class DistributedTest(unittest.TestCase):
             # we need to move the CPU tensor to CUDA in case we are doing cpu_offload.
             shard_loss = shard_loss.cuda()
         shard_state_dict = model.state_dict()
+
+        if config.get("state_dict_on_rank_0_only", False):
+            if torch.distributed.get_rank() != 0:
+                assert shard_state_dict == {}
+                # rank 0 shard_state_dict test covered in the following test.
+                # return is needed here, because with state_dict_on_rank_0_only=True, the following assert will fail on rank!=0
+                return
 
         try:
             torch.testing.assert_allclose(ref_loss.cpu(), shard_loss.cpu())
@@ -264,7 +272,15 @@ class TestMixedPrecision(DistributedTest):
 
     @staticmethod
     def _test_dtypes(
-        cfg: Dict, autocast, in_dtype, p_dtype, loss_dtype, reduce_dtype, rank, group, expected_buffer_type=None
+        cfg: Dict,
+        autocast,
+        in_dtype,
+        p_dtype,
+        loss_dtype,
+        reduce_dtype,
+        rank,
+        group,
+        expected_buffer_type=None,
     ):
         # Patch torch.distributed.reduce_scatter to check the dtype of the reduction
         orig_reduce_scatter = torch.distributed.reduce_scatter
@@ -375,6 +391,13 @@ class TestComparisonToPyTorchDDP(DistributedTest):
         test_fn = functools.partial(self._test_identical_outputs, model_fn, config)
         spawn_and_init(test_fn)
 
+    @parameterized.expand([[True], [False]], name_func=rename_test)
+    def test_state_dict_on_rank_0_only(self, state_dict_on_rank_0_only):
+        config = {"state_dict_on_rank_0_only": state_dict_on_rank_0_only}
+        model_fn = functools.partial(TransformerWithSharedParams)
+        test_fn = functools.partial(self._test_identical_outputs, model_fn, config)
+        spawn_and_init(test_fn)
+
     @parameterized.expand([[{"checkpoint_act": False}], [{"checkpoint_act": True}]], name_func=rename_test)
     def test_mixture_of_experts(self, moe_config):
         fsdp_config = {"mixed_precision": True}
@@ -454,6 +477,24 @@ class TestParamInit(DistributedTest):
         assert not objects_are_equal(ref_output, new_output), "new_output did not reflect change to param after init"
 
 
+class TestReduceScatterProcessGroup(DistributedTest):
+    def test_reduce_scatter_process_group_size(self):
+        """Ensure that reduce_scatter_process_group same size with the world size."""
+        test_fn = functools.partial(self._test_reduce_scatter_process_group_size, config={})
+        spawn_and_init(test_fn, world_sizes=[2])
+
+    @classmethod
+    def _test_reduce_scatter_process_group_size(self, rank, group, config):
+        model = self._get_model(group, config)
+        assert model.process_group_reduce_scatter.size() == model.world_size
+
+    @classmethod
+    def _get_model(self, group, config):
+        with torch.no_grad():  # required for multiprocessing
+            model = NestedWrappedModule(group, wrapper_config=config)
+            return FullyShardedDataParallel(model, group, **config)
+
+
 class TestSerialization(DistributedTest):
     @parameterized.expand([[False, False], [True, False], [True, True], [False, True]], name_func=rename_test)
     def test_pickle(self, mixed_precision, cpu_offload):
@@ -481,6 +522,7 @@ class TestSerialization(DistributedTest):
     def _test_multiprocessing(self, rank, group, config):
         mp = torch.multiprocessing.Pool(1)
         dummy_group = DummyProcessGroup(rank=group.rank(), size=group.size())
+        config["process_group_reduce_scatter"] = DummyProcessGroup(rank=group.rank(), size=group.size())
         model = mp.apply(self._get_model, (dummy_group, config))
         if not config["cpu_offload"]:
             model = model.cuda()
@@ -498,6 +540,7 @@ class TestSerialization(DistributedTest):
         for m in model.modules():
             if isinstance(m, FullyShardedDataParallel):
                 m.process_group = group
+                m.process_group_reduce_scatter = torch.distributed.new_group()
         optim = torch.optim.SGD(model.parameters(), lr=0.01, momentum=0.9)
         input = model.module.get_input(torch.device("cuda"))
         output = model(*input)
@@ -592,10 +635,21 @@ class TestNoGrad(DistributedTest):
         assert objects_are_equal(ref_output, no_grad_output, raise_exception=True)
 
 
-@pytest.mark.skipif(torch_version() < (1, 8, 0), reason="pytorch version >= 1.8.0 required")
+keys = ["flatten_parameters", "ssd_offload"]
+MODULE_CONFIG_OPTIONS = [[dict(zip(keys, config))] for config in itertools.product([True, False], repeat=len(keys))]
+
+
 class TestModuleProperties(DistributedTest):
-    @parameterized.expand([[{"flatten_parameters": False}], [{"flatten_parameters": True}]], name_func=rename_test)
+    @parameterized.expand(MODULE_CONFIG_OPTIONS, name_func=rename_test)
     def test_named_parameters(self, config):
+        if config["ssd_offload"] and not config["flatten_parameters"]:
+            pytest.skip("SSD offload does not support non flattened parameters.")
+        if config["ssd_offload"]:
+            offload_config = OffloadConfig(offload_type="ssd_offload")
+        else:
+            offload_config = None
+        del config["ssd_offload"]
+        config["offload_config"] = offload_config
         test_fn = functools.partial(self._test_named_params, config=config)
         spawn_and_init(test_fn)
 
@@ -616,7 +670,7 @@ class TestModuleProperties(DistributedTest):
             for before_nm, after_nm in zip(before_wrap_params, after_wrap_params):
                 assert before_nm[0] == after_nm[0]
         else:
-            named_params_flat = [p for p in after_wrap_params][0][0]
+            named_params_flat = [p for p in after_wrap_params][0][1]
             assert "flat_param_0" in named_params_flat
 
         # Compare name and size under the `summon_full_params` context.